--- Pipfile
@@ -1,14 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-
-[packages]
-toml = "*"
-cerberus = "*"
-ninja-syntax = "*"
-
-[requires]
-python_version = "3.7"
