+++ pyproject.toml
@@ -1,0 +1,21 @@
+[tool.poetry]
+name = "aim"
+version = "0.1.0"
+description = "Simple C++ build tool."
+authors = ["David Walker <david.walker@email.com>"]
+license = "MIT"
+
+[tool.poetry.dependencies]
+python = "^3.7"
+toml = "^0.10.0"
+cerberus = "^1.3"
+ninja-syntax = "^1.7"
+
+[tool.poetry.dev-dependencies]
+
+[build-system]
+requires = ["poetry>=0.12"]
+build-backend = "poetry.masonry.api"
+
+[tool.poetry.scripts]
+aim = "aim.aim:entry"
