+++ src/aim/gccbuildrules.py
@@ -1,9 +1,9 @@
 from ninja_syntax import Writer
 
-from typedefs import StringList
+from aim.typedefs import StringList
 
 
-class WindowsBuildRules:
+class GCCBuildRules:
     def __init__(self, nfw: Writer, compiler: str):
         self.nfw = nfw
         self.compiler = compiler
@@ -16,10 +16,11 @@ class WindowsBuildRules:
         defines = " ".join(defines)
         includes = " ".join(includes)
 
-        command = f"{self.compiler} {defines} {includes} /showIncludes {flags} -c $in"
+        command = f"{self.compiler} {flags} -MMD -MF deps.d -c {defines} {includes} $in"
         self.nfw.rule(name="compile",
-                      description='Compile source files to object files',
-                      deps="msvc",
+                      description='Compiles source files into object files',
+                      deps="gcc",
+                      depfile="deps.d",
                       command=command)
         self.nfw.newline()
 
@@ -40,9 +41,9 @@ class WindowsBuildRules:
         includes = " ".join(includes)
         linker_args = " ".join(linker_args)
 
-        command = f"{self.compiler} {defines} {flags} {includes} $in /link /out:{exe_name} {linker_args}"
+        command = f"{self.compiler} {defines} {flags} {includes} $in -o {exe_name} {linker_args}"
         self.nfw.rule(name="exe",
-                      description="Build an executable.",
+                      description="Builds an executable.",
                       command=command)
         self.nfw.newline()
 
@@ -57,8 +58,8 @@ class WindowsBuildRules:
         includes = " ".join(includes)
         linker_args = " ".join(linker_args)
 
-        command = f"{self.compiler} {defines} {flags} {includes} $in /link /DLL /out:{lib_name} {linker_args}"
+        command = f"{self.compiler} {defines} -fPIC -shared -fvisibility=hidden {flags} {includes} $in -o {lib_name} {linker_args}"
         self.nfw.rule(name="shared",
-                      description="Build an shared library.",
+                      description="Builds a shared library.",
                       command=command)
         self.nfw.newline()
